+++ src/main/java/ElSilencioKoffee_Backend/inventory/repositories/InventoryRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface InventoryRepository extends JpaRepository<Inventory, Long> {
+public interface InventoryRepository extends JpaRepository<Inventory, Integer> {
 
     Optional<Inventory> findByProductId(Long productId);
 }
