+++ src/main/java/ElSilencioKoffee_Backend/production/repositories/ProductionSectionRepository.java
@@ -3,5 +3,5 @@ package ElSilencioKoffee_Backend.production.repositories;
 import ElSilencioKoffee_Backend.production.entities.ProductionSection;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductionSectionRepository extends JpaRepository<ProductionSection, Long> {
+public interface ProductionSectionRepository extends JpaRepository<ProductionSection, Integer> {
 }
